beta endpoints for appeals
